Update unit tests to cover valid an invalid edge cases
